small changes in shear layer
Former-commit-id: 20d47c04c879712794c9683d376c804069181f4c [formerly b45ea4600ae49441a1a7f3d89abdb78f88b58d88]
Former-commit-id: b004b2633a8f15e77020fff17a2944cbb3c8f3eb

+++ src/examples/step-shearLayer2D/shear_layer.cpp
@@ -71,13 +71,13 @@ int main(int argc, char** argv) {
 
 	double CFL = .4;
 	if (argc >= 7) {
-		CFL = std::atoi(argv[6]);
+		CFL = std::atof(argv[6]);
 	}
 	pout << "... CFL:  " << CFL << endl;
 
 	double stencil_scaling = 1.0;
 	if (argc >= 8) {
-		stencil_scaling = std::atoi(argv[7]);
+		stencil_scaling = std::atof(argv[7]);
 	}
 	pout << "... stencil_scaling:  " << stencil_scaling << endl;
